refactor the user service and change association and responsibilty to better fit Package By Layer
