+++ ComfortCare.Data/Interfaces/IUserRepo.cs
@@ -5,6 +5,6 @@ namespace ComfortCare.Data.Interfaces
     public interface IUserRepo
     {
         public bool ValidateUserExist(string username, string password);
-        public EmployeeSchemaModel GetUsersWorkSchedule(string username, string password);
+        public Employee GetUsersWorkSchedule(string username, string password);
     }
 }
