+++ ComfortCare.Service/Models/EmployeeSchemaModel.cs
@@ -1,0 +1,8 @@
+﻿namespace ComfortCare.Service.Models
+{
+    public class EmployeeSchemaModel
+    {
+        public string Name { get; set; }
+        public List<AssignmentSchemaModel> Assignments { get; set; }
+    }    
+}
